requirements.txt
flask>=3
waitress>=3
requests>=2.31
paho-mqtt>=2.0,<3
PyYAML>=6
